api-python/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
psycopg2-binary==2.9.11